ml_assignment/image_classification/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

